// src/app/models/store-user.model.ts
import { StoreDetails } from './store-details.model';

export class StoreUser {
    AUTH_TOKEN: any;
    VALID_TILL: any;
    user_id: number;
    store_id: boolean;
    first_name: any;
    last_name: any;
    email: any;
    image:any;
    email_verified: any;
    enabled: any;
    dob: any;
    contact_number: any;
    storeDetails: StoreDetails;
  }
  